CE 675 Project 1 updated
This is an update to the solution for the first CE project. The
equations are now formulated correctly (hopefully) and the mesh
generation is switched back to using two variables. It still
doesn't converge on a reasonable solution, but that's likely
because the solver doesn't implement any sort of arc length
constraint.