Added an implicit time integrator.
This adds a basic time integration algorithm that works for both
trapazoid rule and backward difference stuff. It solves the problem the
explicit integrator had with natural boundary conditions.

Additionally, this separates out the heat equation stuff into a separate
file so that the same stuff can be used for both the console version and
the gui version.